pyproject.toml
[tool.poetry]
name = "zagger"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.3.2"
python-dotenv = "^1.0.0"
replicate = "^0.22.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
